rename: RaceGameCalculator -> RaceCalculator 이름 변경
